Update README.md
Welcome to Joke Battle, where wit meets whimsy in a high-stakes clash of comedy! 🎭🤣 Ever wondered who could make a rubber chicken laugh? Or which joke could bring a smile to a statue? This is your chance to prove it!

In this game, two fearless jokesters face off with their best (and possibly worst) jokes. With a sprinkle of React magic and a dash of Node.js wizardry, you’ll submit your top-notch quips and let our AI decide whose humor reigns supreme. Will you be crowned the King or Queen of Comedy, or will your jokes fall flatter than a pancake at a circus?

Get ready to laugh, cringe, and maybe shed a tear or two (of joy or laughter, we hope)! May the funniest joke win, and may the best comedian triumph in the battle of giggles!